fix: fix merging error

+++ reverse-auction/src/main/java/vn/edu/hcmuaf/reverseauction/dto/AuctionResponse.java
@@ -10,7 +10,7 @@ public record AuctionResponse(
         String description,
         BigDecimal budgetMax,
         String status,
-        Long currentLowestPrice,
+        BigDecimal currentLowestPrice,
         LocalDateTime endDate,
         List<BidResponse> bids
 ) {
